fix(#28): move configuration functions to module
